Allow running tests w/o psiphon
Otherwise the life of a contributor that doesn't have access to
the Psiphon secret would be miserable.
